swarm/pss: Remove go1.7 test workaround
